Few spelling/grammar corrections (#17)
Huge thanks to Rich & PagerDuty for publishing this document. Have been aware of most of the topic areas covered, but this brought accessible language, detailed how-to instructions, and diverse external resources for deeper learning.

Much care was taken to keep the content neutral to product/provider/service. Great job! One small improvement to this area is for representing the browser (such as on slide #166). Here is one place I know of to find icons with an open source license: https://thenounproject.com/search/?q=web%20browser

Also the "smaller text with an arrow" technique is effective (and hilarious) so I'm stealing it for one of my future presentations :smiling_imp: